parser/requirements.txt
langchain_community
PyMuPDF
langchain-core
langchain-openai
python-dotenv
